fix: deleting proxies
